move convertToLongFormat to util file
